fixes #51: Add new offline setting: redirect to chatroom
This introduces a new configurable 'offline setting': redirect to a chat room, adding this option to the database and the admin console.

The embedded converse client is modified to no longer hardcoded falling back to a chatroom, but to do this only when the offline settings instruct it to do so. This routine needs a lot more work (see #49).