fixes #59: Additional permissions for XStream
In a later version of XStream, explicit permission needs to be given for classes to be (de)serialized as XML.

As in older versions, these permissions were not needed, they're absent in the code.

This commit adds those permissions, building on an earlier commit that introduced the first few permissions.